+++ CHANGELOG.md
@@ -19,6 +19,7 @@ Changelog
 * Directly open Xposed pages, thanks @[liudongmiao](https://github.com/liudongmiao)
 * Added multiple template definitions
 	* The first template will always be applied to new applications
+* Added allow contacts by group
 * Updated simplified Chinese translation
 * Updated traditional Chinese translation